Move Types.tsx
